requirements.txt
Flask==3.1.2
flask-cors==3.0.10
pymongo==4.15.5
python-dotenv==1.2.1
langchain
langchain-google-genai
oci
Pillow